Improve Pi Non-GPU-accel exposure
Change exposure setters to accept a floating point input (rather than integer)
Update the UI to change exposure in increments of 0.1 (rather than 1.0)
Update quirky PI camera logic to use the raw_ interface with scaling/offset logic matching the GPU-accelerated pi3 camera logic from Delcan.
Added logic to disable auto-white-balance in the PI camera, which should yeild more consistent vision processing results.

+++ photon-server/src/main/java/org/photonvision/vision/camera/FileVisionSource.java
@@ -89,7 +89,7 @@ public class FileVisionSource implements VisionSource {
         }
 
         @Override
-        public void setExposure(int exposure) {}
+        public void setExposure(double exposure) {}
 
         @Override
         public void setBrightness(int brightness) {}
